chore: moved object nesting models into the cdk library
