+++ libs/cdk/src/lib/utils/datatypes/index.ts
@@ -1,3 +1,4 @@
-export * from './hash.datatype';
-export * from './size.datatype';
 export * from './color-accent.datatype';
+export * from './hash.datatype';
+export * from './object-nesting.types';
+export * from './size.datatype';
